+++ lib/views/function_screens/capitalizacion_screen.dart
