+++ lib/views/function_screens/series_variables_screen.dart
