+++ lib/views/home_screens/home_screen.dart
@@ -1,12 +1,12 @@
 import 'package:finanpro_v2/controllers/text_formater.dart';
-import 'package:finanpro_v2/views/screens/amortizacion_screen.dart';
-import 'package:finanpro_v2/views/screens/anualidades_screen.dart';
-import 'package:finanpro_v2/views/screens/capitalizacion_screen.dart';
-import 'package:finanpro_v2/views/screens/gradientes_screen.dart';
-import 'package:finanpro_v2/views/screens/interes_compuesto_screen.dart';
-import 'package:finanpro_v2/views/screens/interes_simple_screen.dart';
-import 'package:finanpro_v2/views/screens/series_variables_screen.dart';
-import 'package:finanpro_v2/views/screens/tir_screen.dart';
+import 'package:finanpro_v2/views/function_screens/amortizacion_screen.dart';
+import 'package:finanpro_v2/views/function_screens/anualidades_screen.dart';
+import 'package:finanpro_v2/views/function_screens/capitalizacion_screen.dart';
+import 'package:finanpro_v2/views/function_screens/gradientes_screen.dart';
+import 'package:finanpro_v2/views/function_screens/interes_compuesto_screen.dart';
+import 'package:finanpro_v2/views/function_screens/interes_simple_screen.dart';
+import 'package:finanpro_v2/views/function_screens/series_variables_screen.dart';
+import 'package:finanpro_v2/views/function_screens/tir_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:finanpro_v2/controllers/auth_controller.dart';
